requirements.txt
# Core dependencies
PyQt6>=6.5.0
matplotlib>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0

# Optional dependencies for enhanced functionality
plotly>=5.14.0
pyqtgraph>=0.13.3
pillow>=9.5.0  # For image handling

# Packaging tools
pyinstaller>=5.10.0
